Brings back checks for older MESA versions.
Reverts commits 6e65547 and 479df8e.
These commits caused the build to fail on DragonFlyBSD.

mesa-libs 21.3.9 is not by any means ancient, it is only 3 years old and
it is still receiving critical bug fixes, and most importantly it is
still being used on DragonFlyBSD, also these are relatively small and
harmless checks.

Signed-off-by: b-aaz <b-aazbsd.proton.me>